Create ui.R
Like most R Shiny app, you have to include both the UI and Server codes in one folder (named "PK_Sim" in my case), and set the working directory where the folder is located.

Type in shiny::runApp('PK_Sim') in the console of the RStudio to start.

This app simulates the PK or both 2-compartment oral and 1-compartment IV scenarios. It also allows you to adjust the clearance, volume, ka, and dosing changes and lets you see the concentration simulated in real time. You can also compare the difference between oral and IV intake by toggling the checkboxes on the sidebar.